Add support for Delta3, improve typing for protobuf fields (#12)
This change adds support for Delta 3 (tested of Delta 3 Plus) and does some refactoring of fields that makes them tied to typing stubs and reduces chance of misspelling or using the wrong type and makes them easier to find and define.

It also fixes generated protobufs for River 3 - previous version does not have a namespace which made collisions with Delta 3 protocol buffers.

* Refactor protobuf fields and make their typing easier
* Fix generated protos not having correct package, add typing stubs
* Add support for Delta 3, use better icons for sensors
* Clean up and merge D3 sensors and their names
* Add missing newline to icons.json

---------

Co-authored-by: GnoX <dev@gnox.dev>